Run calibrator on separate thread (nonblocking UI)
